requirements.txt
requests
colorama
rich.console